@@ -1,8 +1,8 @@
 # Be sure to restart your server when you modify this file.
 
-RailsTemplate::Application.config.session_store :cookie_store, :key => '_rails_template_session'
+TEMPLATE::Application.config.session_store :cookie_store, key: '_TEMPLATE_session'
 
 # Use the database for sessions instead of the cookie-based default,
 # which shouldn't be used to store highly confidential information
 # (create the session table with "rails generate session_migration")
-# RailsTemplate::Application.config.session_store :active_record_store
+# TEMPLATE::Application.config.session_store :active_record_store
