@@ -4,4 +4,4 @@
 # If you change this key, all old signed cookies will become invalid!
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-RailsTemplate::Application.config.secret_token = 'd32edb255c5e401c1b1610c0ffdbad83bcdd6768a3d4670baa996471b7e0fe9a3628a6885914e5c0e59f5077b55a9ca976fff16918613d091bba40d43c05b0e5'
+TEMPLATE::Application.config.secret_token = nil
